+++ packages/helpers/index.ts
@@ -1,11 +1,2 @@
-import { palette } from "@guardian/src-foundations"
-
-const storybookBackgrounds = {
-	light: { name: "light", value: palette.neutral[100] },
-	blue: { name: "blue", value: palette.brand.main },
-	dark: { name: "dark", value: palette.neutral[10] },
-}
-
-Object.freeze(storybookBackgrounds)
-
-export { storybookBackgrounds }
+export { storybookBackgrounds } from "./storybook-bg"
+export { Toggle } from "./storybook-bg-toggle"
