+++ packages/helpers/storybook-bg.ts
@@ -1,0 +1,11 @@
+import { palette } from "@guardian/src-foundations"
+
+const storybookBackgrounds = {
+	light: { name: "light", value: palette.neutral[100] },
+	blue: { name: "blue", value: palette.brand.main },
+	dark: { name: "dark", value: palette.neutral[10] },
+}
+
+Object.freeze(storybookBackgrounds)
+
+export { storybookBackgrounds }
